+++ lib/hypertrack.dart
@@ -3,14 +3,16 @@ import 'dart:ffi';
 import 'package:flutter/services.dart';
 import 'package:hypertrack_plugin/data_types/json.dart';
 import 'package:hypertrack_plugin/data_types/location_error.dart';
+import 'package:hypertrack_plugin/data_types/location_with_deviation.dart';
 import 'package:hypertrack_plugin/data_types/result.dart';
 import 'package:hypertrack_plugin/src/sdk_method.dart';
 import 'package:hypertrack_plugin/src/serialization/availability.dart';
 import 'package:hypertrack_plugin/src/serialization/device_id.dart';
 import 'package:hypertrack_plugin/src/serialization/device_name.dart';
-import 'package:hypertrack_plugin/src/serialization/geotag.dart';
+import 'package:hypertrack_plugin/src/serialization/geotag_data.dart';
 import 'package:hypertrack_plugin/src/serialization/hypertrack_error.dart';
 import 'package:hypertrack_plugin/src/serialization/location_result.dart';
+import 'package:hypertrack_plugin/src/serialization/location_with_deviation_result.dart';
 import 'package:hypertrack_plugin/src/serialization/metadata.dart';
 import 'package:hypertrack_plugin/src/serialization/tracking_state.dart';
 
@@ -27,13 +29,16 @@ class HyperTrack {
     bool? requireBackgroundTrackingPermission,
     bool? loggingEnabled,
     bool? allowMockLocations,
+    bool? automaticallyRequestPermissions,
   }) {
     return _invokeSdkVoidMethod(SdkMethod.initialize, {
       _keyPublishableKey: publishableKey,
       _keyRequireBackgroundTrackingPermission:
           requireBackgroundTrackingPermission ??= false,
       _keyLoggingEnabled: loggingEnabled ??= false,
-      _keyAllowMockLocations: allowMockLocations ??= false
+      _keyAllowMockLocations: allowMockLocations ??= false,
+      _keyAutomaticallyRequestPermissions: automaticallyRequestPermissions ??=
+          true,
     }).then((value) => HyperTrack._());
   }
 
@@ -70,12 +75,11 @@ class HyperTrack {
         SdkMethod.setAvailability, serializeAvailability(available));
   }
 
-  
   /// Sets the name for the device
   void setName(String name) {
     _invokeSdkVoidMethod(SdkMethod.setName, serializeDeviceName(name));
   }
-  
+
   /// Sets the metadata for the device
   void setMetadata(JSONObject data) {
     _invokeSdkVoidMethod(SdkMethod.setMetadata, serializeMetadata(data));
@@ -88,9 +92,21 @@ class HyperTrack {
 
   /// Adds a new geotag
   Future<Result<Location, LocationError>> addGeotag(JSONObject data) {
-    return _invokeSdkMethod(SdkMethod.addGeotag, serializeGeotag(data))
+    return _invokeSdkMethod(
+            SdkMethod.addGeotag, serializeGeotagData(data, null))
         .then((value) {
       return deserializeLocationResult(value);
+    });
+  }
+
+  /// Adds a new geotag with expected location
+  Future<Result<LocationWithDeviation, LocationError>>
+      addGeotagWithExpectedLocation(
+          JSONObject data, Location expectedLocation) {
+    return _invokeSdkMethod(
+            SdkMethod.addGeotag, serializeGeotagData(data, expectedLocation))
+        .then((value) {
+      return deserializeLocationWithDeviationResult(value);
     });
   }
 
@@ -99,7 +115,7 @@ class HyperTrack {
     return _invokeSdkMethod(SdkMethod.getLocation).then((value) {
       return deserializeLocationResult(value);
     });
-  }  
+  }
 
   /// Subscribe to tracking intent changes
   Stream<bool> get onTrackingChanged {
@@ -159,6 +175,7 @@ const _keyRequireBackgroundTrackingPermission =
     'requireBackgroundTrackingPermission';
 const _keyLoggingEnabled = 'loggingEnabled';
 const _keyAllowMockLocations = 'allowMockLocations';
+const _keyAutomaticallyRequestPermissions = 'automaticallyRequestPermissions';
 
 // channel for invoking SDK methods
 const _methodChannel = MethodChannel('$_channelPrefix/methods');
@@ -168,8 +185,7 @@ const EventChannel _trackingStateEventChannel =
     EventChannel('$_channelPrefix/tracking');
 const EventChannel _availabilityEventChannel =
     EventChannel('$_channelPrefix/availability');
-const EventChannel _errorEventChannel =
-    EventChannel('$_channelPrefix/errors');
+const EventChannel _errorEventChannel = EventChannel('$_channelPrefix/errors');
 
 Future<T> _invokeSdkMethod<T>(SdkMethod method, [dynamic arguments]) {
   return _methodChannel.invokeMethod(method.name, arguments).then((value) {