+++ lib/src/serialization/location_result.dart
@@ -2,40 +2,21 @@ import 'package:hypertrack_plugin/data_types/location_error.dart';
 import 'package:hypertrack_plugin/data_types/result.dart';
 import 'package:hypertrack_plugin/src/serialization/common.dart';
 import 'package:hypertrack_plugin/src/serialization/location_error.dart';
+import 'package:hypertrack_plugin/src/serialization/result.dart';
 
 import '../../data_types/hypertrack_error.dart';
 import '../../data_types/location.dart';
+import 'location.dart';
 
 Result<Location, LocationError> deserializeLocationResult(
     Map<Object?, Object?> response) {
-  try {
-    Map<String, dynamic> data = response.cast<String, dynamic>();
-    switch (data[keyType]) {
-      case _typeSuccess:
-        return Result.success(deserializeLocation(data[keyValue]));
-      case _typeFailure:
-        return Result.error(deserializeLocationError(data[keyValue]));
-      default:
-        throw Exception("Invalid location response: ${response}");
-    }
-  } catch (e) {
-    throw Exception("Invalid location response: ${response} $e");
+  Result<Map<Object?, Object?>, Map<Object?, Object?>> result =
+      deserializeResult(response);
+  if (result is Success<Map<Object?, Object?>, Map<Object?, Object?>>) {
+    return Result.success(deserializeLocation(result.value));
+  } else if (result is Error<Map<Object?, Object?>, Map<Object?, Object?>>) {
+    return Result.error(deserializeLocationError(result.error));
+  } else {
+    throw Exception("Illegal argument: ${response}");
   }
 }
-
-Location deserializeLocation(Map<Object?, Object?> map) {
-  try {
-    final data = (map[keyValue] as Map<Object?, Object?>).cast<String, double>();
-    return Location(
-        data[_keyLatitude]!,
-        data[_keyLongitude]!
-    );
-  } catch(e) {
-    throw Exception("Invalid location $map $e");
-  }
-}
-
-const _typeSuccess = "success";
-const _typeFailure = "failure";
-const _keyLatitude = "latitude";
-const _keyLongitude = "longitude";