+++ test/json_test.dart
@@ -2,23 +2,18 @@ import 'package:hypertrack_plugin/data_types/json.dart';
 import 'package:test/test.dart';
 
 void main() {
-  test('Serialize-deserialize JSON', () {
+  test('Serialize JSON', () {
     JSONObject testValue = JSONObject({
-      "null": null,
+      "null": JSONNull(),
       "string": JSONString("value"),
       "number": JSONNumber(1),
       "bool": JSONBool(false),
-      "object": JSONObject({
-        "data": JSONBool(true)
-      }),
-      "intArray": JSONArray([JSONNumber(1), JSONNumber(2)]),
+      "object": JSONObject({"data": JSONBool(true), "value": JSONNull()}),
+      "nullArray": JSONArray([JSONNull(), JSONNull()]),
+      "intArray": JSONArray([JSONNumber(1), JSONNumber(2), JSONNull()]),
       "objectArray": JSONArray([
-        JSONObject({
-          "index": JSONNumber(0)
-        }),
-        JSONObject({
-          "index": JSONNumber(1)
-        })
+        JSONObject({"index": JSONNumber(0)}),
+        JSONObject({"index": JSONNumber(1)})
       ])
     });
 
@@ -28,17 +23,12 @@ void main() {
       "string": "value",
       "number": 1,
       "bool": false,
-      "object": {
-        "data": true
-      },
-      "intArray": [1,2],
+      "object": {"data": true, "value": null},
+      "nullArray": [null, null],
+      "intArray": [1, 2, null],
       "objectArray": [
-        {
-          "index": 0
-        },
-        {
-          "index": 1
-        }
+        {"index": 0},
+        {"index": 1},
       ]
     };